requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.5
# webrtcvad の Windows 向けビルド済みバイナリ版（import webrtcvad で使用可）
webrtcvad-wheels
